[refactor] create validationUtils file containing input validation functions
