+++ src/screens/scanning/ManualVoucherScreen.tsx
@@ -1,5 +1,4 @@
 import React, { useState } from 'react';
-import { z } from 'zod';
 import { ButtonMagenta } from '../../../assets/Components';
 import {
   ButtonTextWhite,
@@ -18,6 +17,10 @@ import {
 } from './styles';
 import InputField from '../../components/InputField/InputField';
 import StandardLogo from '../../components/common/StandardLogo';
+import {
+  validateSerialNumberInput,
+  validateVoucherAmount,
+} from '../../utils/validationUtils';
 
 function ManualVoucherScreen() {
   const [transactionID, setID] = useState<string>('');
@@ -39,41 +42,6 @@ function ManualVoucherScreen() {
   const onChangeVoucherAmount = (text: string) => {
     const value = text.replace(/[^\d.]/g, '');
     setVoucherAmount(value);
-  };
-
-  const validateSerialNumberInput = (input: string) => {
-    try {
-      const SNSchema = z.coerce.number().gt(0);
-      SNSchema.parse(input);
-    } catch (error) {
-      // eslint-disable-next-line no-console
-      console.log(error);
-    }
-  };
-
-  const validateCurrencyFormat = (input: string) => {
-    // confirms that any input has either 0 or 2 decimal digits
-    const inputParts = input.split('.');
-    if (inputParts.length > 1 && inputParts[1].length !== 2) {
-      return false; // falsy value raises the error in zod
-    }
-    return true;
-  };
-
-  const validateVoucherAmount = (input: string) => {
-    try {
-      const currencySchema = z.coerce
-        .number() // ensures that input is a valid number
-        .gt(0)
-        .lte(10) // less than or equal to 10 dollars
-        .refine(() => validateCurrencyFormat(input), {
-          message: 'Input must be in a valid currency format',
-        });
-      currencySchema.parse(input);
-    } catch (error) {
-      // eslint-disable-next-line no-console
-      console.log(error);
-    }
   };
 
   return (
